subtask(F-3-2-tui-interface-client-boundary): Move TUI HTTP client code out of Client Integrations
