+++ crates/tanren-tui-app/Cargo.toml
@@ -14,9 +14,13 @@ workspace = true
 [dependencies]
 crossterm = { workspace = true }
 ratatui = { workspace = true }
+reqwest = { workspace = true, features = ["cookies", "json"] }
 secrecy = { workspace = true }
-tanren-client-integrations = { path = "../tanren-client-integrations" }
 tanren-contract = { path = "../tanren-contract" }
+tanren-identity-policy = { path = "../tanren-identity-policy" }
+serde = { workspace = true }
+serde_json = { workspace = true }
 thiserror = { workspace = true }
 tokio = { workspace = true }
 tracing = { workspace = true }
+uuid = { workspace = true }
